kafka-python-camera-stream-consumer/requirements/requirements.txt
kafka-python
opencv-python
Flask
pymongo